[in-progress] separate mmu tables from ept tables when caching.
This fixes a bug I've been looking for for a week. The EPT and the MMU
tables are largely similar, so there is a shared "table" code between
them. Additionally, each table level in the page tables has a struct for
it, which contains information for tracking child tables and pages. They
attempt to cache their memory -- that is, the 4K allocated region for
each table is kept for reused in the next table.

However, these tables use different _physical_ address types -- the EPT
uses real physical addresses, while the MMU uses guest-physical
addresses. Tables, before now, were put into the same cache pool,
because I'm dumb. This meant that an MMU table could be reused in the
EPT, and, if the table had been "realized" (a 4K page allocated for it)
it's physical address would be invalid (because it'd be a guest-physical
address).

The fix is to maintain two caches, one for EPT tables and one for MMU
tables. That took way longer to figure out than it should have. It
smelled so strongly like invalidation issues in the way it manifested,
though, that I ended up pursuing the wrong thing for days. Oh well. :)