Add basic unwinding for userspace threads in kernel.
This is meant as a very simple debugging tool, to trace a thread's call
stack before it entered the kernel. It cannot resolve symbols, it just
prints instruction pointers that it finds in the call frames.

Currently it blindly reads memory, so in the future we should add checks
for validity of mappings, otherwise this is a major vuln.

This function is not intended for general use.