Speed up object copy.
When a given range isn't found, traverse the range tree for the next
range, skipping all pages in between. This dramatically speeds up object
copy when copying sets of pages that are not backed by anything.

We will need to handle the case where we don't necessarily know if those
pages are not backed by something, as will be the case with paging in,
but we can still apply an optimization there where the pager can inform
the system that a range has no backing pages.