Reduce sizeof(struct page) to 16.
struct page now stores flags inside the low bits of the address,
requiring new utility functions to extract the address or the flags.
Also, the struct list is 16 bytes long but we don't need a full linked
list for our page lists, so we'll just manually implement a simple
single linked list to track struct pages. The extra code is worth the
memory savings because of how many struct pages there could be.