Fix toolchain bootstrap for crt files.
When rust is bootstrapping, it does some Weird Stuff (tm) with sysroot,
which plays weirdly with how the toolchain works (basically useless in a
bootstrapping env unless you specify a correct sysroot). As a result, it
tries to link rust libs with _host_ *crt*.o files, which is Bad News.
This only manifests as an actual error, however, on Ubuntu (or similar)
systems that do not have /usr/lib/*crt*.o files (they put them
elsewhere), since it would cause a linker error. On systems with those
files, it would silently link against host crt files, which is not
ideal, but since they'd get overwritten after the bootstrapping phase,
it was okay.

Instead, now, we copy in the crt files into the toolchain's library path
(lib/clang/<version>) which is where crtbegin and friends are. So that
works okay. The downside is that the sysroot versions of the crt files
will get ignored, as this lib path is searched first. We could fix this
in the future, but since we rarely (if ever) modify crt files, this is
okay for now.