Support non-GB aligned kernel-object-space size.
The kernel allocates a section of the object space (or guest-phys)
memory for itself, which is shared by all object spaces (similar to how
this is done for virtual memory). Twizzler now supports object space
sizes that, when divided by two, are not GB-aligned (support for x86
CPUs this way).