Update gs segment to refer to current thread repr.
[in-progress -- does not run]

Previously, gs was essentially hardwired to TWZSLOT_THRD. But since now
we're supporting the thread repr being in arbitrary locations, we need
to actually store that location in gs so the thread can get to it. This
requires any previous functions that called become or thread_spawn to
include updates to the gs segment register.

This is a partial update. We'll need to still do this properly for twix
and twix v2, along with updating twix to handle thread-local twix_conn
(or, at least get this working properly).